Changed multiple use cases for important scripts
All subclasses of waifu are now legacy, they can be ignored and/or archived. The main Waifu.h header will function as the waifu use case for any and all waifu creation functions. Generally speaking, the architechture has not changed all that much, implicit functions were edited to take in ONLY SFML based datatypes to minimize compiler errors. The biggest change of this update is the game flow, it now goes from main menu to GamePlay state. The functions to make the game playable are still lakcluster and unfinished, also the main waifu is not in the correct place. Future versions will elicit the data to persist between states, but since the main program and data for each waifu is present in the main, this should not be a problem. - Gian